Fix: hooks error

+++ hooks.init.sh
@@ -1,6 +1,12 @@
 #!/usr/bin/env bash
 
 chmod +x ./hooks/pre-commit.sh
-ln ./hooks/pre-commit.sh ./.git/hooks/pre-commit
+
+if ! [[ -f "./.git/hooks/pre-commit" ]]; then
+  ln ./hooks/pre-commit.sh ./.git/hooks/pre-commit
+fi
 chmod +x ./hooks/commit-msg.sh
-ln ./hooks/commit-msg.sh ./.git/hooks/commit-msg
+if ! [[ -f "./.git/hooks/commit-msg" ]]; then
+  ln ./hooks/commit-msg.sh ./.git/hooks/commit-msg
+
+fi
